requirements.txt
aiogram==3.24.0
aiohttp==3.13.3
aiofiles==25.1.0
matplotlib==3.10.8
requests==2.32.5
python-dotenv==1.2.1
boto3==1.42.30
yandexcloud==0.373.0
